For constant delays, make ldi instructions explicit
Previously, the loading of the constant loops value was left to the
compiler, but in some cases it would load to a different register and
then move it to the right register (in some cases, it would even move
twice).

It turns you can do the loading explicitly, and pass the number of loops
to the assembler as an integer constant. This significantly improves the
accuracy of the delay.