Handle (expected) overflow of the measurement timer
Sometimes, an implementation error causes an overflow of the delay loop
variable, which typically causes the actual delay to be off by a
multiple of 2^16 cycles. Since the measurement timer overflows by the
same amount, such an error was not previously detected. Now, an overflow
of the measurement timer is detected and flagged, except when the number
of cycles is more than 2^16 and the measurement timer is expected to
overflow. In the latter case, the measurement is not totally accurate,
since there is no way to detect how often the measurement timer has
overflowed.