Fix non-constant delays
In these implementation, the overhead was counted in cycles, as noted
down in the comments. However, the microseconds calculated from these
cycle counts were four times too high, presumably because they were
interpreted as loop counts instead of cycle counts. Since the loop count
was compensated using this too high value, almost all delays ended up
too small (by a fixed amount, varying from 1.5us at 16Mhz to over 20us
at 1Mhz).

This commit fixes the overhead calculation, sometimes adding some
additional tweaks where the overhead was not an integer number of
microseconds.

Additionally, in the 1Mhz and 2Mhz implementations it could happen that
the division happening there would result in a zero loop count. Since
the loop starts by decrementing the loop counter, it would underflow and
the loop would run 65536 times instead of 0, resulting in a very big
delay. To fix this, an additional check is done on the loop count to
bail out if it is zero.