Compensate for rounding errors at 2Mhz
There is a /2 division (shift right) in the loop calculation, which
rounds its result, making the delay 2 cycles (half a loop) too short.
This compensates by checking the carry bit after the shift (which
contains the bit shifted out), and introduces two extra branches for two
additional cycles if it is set.

This makes all delays of 7us and bigger perfect, at the expense of a few
additional error cycles below 7us.